requirements.txt
aws-cdk-lib==2.155.0
constructs>=10.0.0,<11.0.0
boto3==1.36.16
fastapi==0.115.8
python-dotenv==1.0.1
uvicorn==0.34.0
pytest==8.3.4
smolagents==1.7.0
openai>=1.55.3
smolagents[litellm]   
streamlit==1.42.0
requests==2.32.3